refactor: rename analysis workflow execution module
